requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
discord.py==2.5.2
frozenlist==1.6.2
fugashi==1.5.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
unidic-lite==1.0.8
yarl==1.20.0
